+++ frontend/store.ts
@@ -1,5 +1,19 @@
 import { createStore, produce } from "solid-js/store";
+import { createResource } from "solid-js";
 import type { Message } from "../types/message";
+
+export type InstanceData = {
+  id: string;
+  url: string;
+};
+
+const [instances, { refetch: refetchInstances }] = createResource<
+  InstanceData[]
+>(async () => {
+  return (await fetch(`/instances`)).json();
+});
+
+export { instances, refetchInstances };
 
 const [state, setState] = createStore<Message[]>([]);
 
@@ -11,26 +25,33 @@ export function addMessage(message: Message): void {
   );
 }
 
-export function sent(clientId: string): number {
+export function sent(instanceId: string): number {
   let result = 0;
   for (const message of state) {
-    if (message.type === "sent" && message.clientId === clientId) {
+    if (message.type === "sent" && message.instanceId === instanceId) {
       result++;
     }
   }
   return result;
 }
 
-export function received(clientId: string): number {
+export function received(instanceId: string): number {
   let result = 0;
   for (const message of state) {
-    if (message.type === "received" && message.clientId === clientId) {
+    if (message.type === "received" && message.instanceId === instanceId) {
       result++;
     }
   }
   return result;
 }
 
-export function getMessages(): Message[] {
-  return state;
+export function getMessages(instanceId: string, type: string): Message[] {
+  if (instanceId == null && type == null) {
+    return state;
+  }
+  return state.filter(
+    (message) =>
+      (instanceId == null || message.instanceId === instanceId) &&
+      (type == null || message.type === type)
+  );
 }